enhance: move the cgo code in the pkg dir to interal dir
