+++ internal/util/cgoconverter/bytes_converter.go
