+++ internal/util/cgoconverter/bytes_converter_test.go
